// src/types/viacep.types.ts
export interface IViaCepData {
  cep: string;
  logradouro: string;
  complemento: string;
  bairro: string;
  localidade: string;
  uf: string;
  ibge: string;
  gia: string;
  ddd: string;
  siafi: string;
}

export interface HttpError extends Error {
  status?: number;
  code?: string;
  message: string;
}